Implementation of functions is done...some schema needs to be changed slightly
Ui class remains to be implemented

i will try to implement it by end of day

+++ src/cartProduct.java
@@ -1,3 +1,5 @@
+
+
 public class cartProduct {
 
     private int cartProductId;
@@ -13,10 +15,11 @@ public class cartProduct {
     }
 
     public double getCartProductPrice() {
+
         return cartProductPrice;
     }
 
-    public void setCartProductPrice(int cartProductPrice) {
+    public void setCartProductPrice(double cartProductPrice) {
 
         this.cartProductPrice = cartProductPrice;
     }
